[future][Connectors-v2] support Mysql8.4+ for mysql-cdc
